refactor: move note_service
