+++ lib/remote_service/remote_service.dart
@@ -1,4 +1,4 @@
 export 'package:simple_notes_app/remote_service/src/authentication/sign_in_remote_service.dart';
 export 'package:simple_notes_app/remote_service/src/authentication/user_account_remote_service.dart';
-export 'package:simple_notes_app/remote_service/src/notes_service.dart';
+export 'package:simple_notes_app/remote_service/src/note/notes_service.dart';
 export 'package:simple_notes_app/remote_service/src/registration/sign_up_remote_service.dart';
